\name{GNAR}
\alias{GNAR}

\title{GNAR package}
\description{A package to fit, predict, and simulate time series using the Generalised Network AutoRegressive (GNAR) model. The main functions are \link[GNAR]{GNARfit}, which fits the model to a time series and network(s), S3 method \link[GNAR]{predict.GNARfit} which predicts from a fitted GNAR model, and \link[GNAR]{GNARsim} which simulates from a GNAR model with specified parameters. For details of the model, see \link[GNAR]{GNARfit}. The package also contains an example network time series in data file \link[GNAR]{fiveNode}, with network \link[GNAR]{fiveNet}, and simulated time series \link[GNAR]{fiveVTS}.}
\references{Knight, M.I., Nunes, M.A. and Nason, G.P. Modelling, detrending and decorrelation of network time series.
	\href{https://arxiv.org/abs/1603.03221}{arXiv preprint}.
}
